+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## [1.0.4](https://github.com/MarcaDian/morphe-patches-yavot/compare/v1.0.3...v1.0.4) (2026-08-09)
+
+### 🐛 Bug Fixes
+
+* Add missing Russian translation for the live-stream error message ([1caed27](https://github.com/MarcaDian/morphe-patches-yavot/commit/1caed27290a0724ab187a034a19f8a338de427e0))
+* Correct string keys for the button countdown timer text ([2473aec](https://github.com/MarcaDian/morphe-patches-yavot/commit/2473aecb72bec193a11e2943924119aebde13cb3))
+* Remove unused strings and show summaries for VoT/live voices switches ([0cd2c86](https://github.com/MarcaDian/morphe-patches-yavot/commit/0cd2c868083dadd69144bb48aaa3f9d532daa800))
+* Resolve IDE warnings ([f0a7b85](https://github.com/MarcaDian/morphe-patches-yavot/commit/f0a7b8564a50122e2a2e20241ac09d452962d00e))
+
+## [1.0.3](https://github.com/MarcaDian/morphe-patches-yavot/compare/v1.0.2...v1.0.3) (2026-08-09)
+
+### 🐛 Bug Fixes
+
+* Stop the built-in voice-over translation when Yandex VoT starts, and vice versa ([498a21d](https://github.com/MarcaDian/morphe-patches-yavot/commit/498a21ded5542ebad4820619de5bd8be16886a73))
+
 ## [1.0.2](https://github.com/MarcaDian/morphe-patches-yavot/compare/v1.0.1...v1.0.2) (2026-08-07)
 
 ### 🐛 Bug Fixes